+++ services/assistant_api/nutrition/__init__.py
@@ -1,0 +1,5 @@
+"""
+Author: L. Saetta
+Date Modified: 2026-05-12
+License: MIT
+"""